pyproject.toml
# SPDX-FileCopyrightText: 2021-2022 Carnegie Mellon University
# SPDX-License-Identifier: 0BSD

[tool.poetry]
name = "sinfonia"
version = "0.2.24.post.dev0"
description = "Tier 1 and 2 components of the Sinfonia system"
authors = [
    "Carnegie Mellon University <satya+group@cs.cmu.edu>",
    "Jan Harkes <jaharkes@cs.cmu.edu>",
]
license = "MIT"
packages = [
    { include = "sinfonia", from = "src" },
    { include = "tests", format="sdist" },
]

[tool.poetry.dependencies]
python = "^3.7"
attrs = "^21.2.0"
click = "^8.0.3"
connexion = {version = "^2.9.0", extras = ["swagger-ui"]}
Flask-APScheduler = "^1.12.3"
Flask-Executor = "^0.10.0"
geopy = "^2.2.0"
jsonschema = {version = "^4.4.0", extras = ["format_nongpl"]}
maxminddb = "^2.2.0"
maxminddb-geolite2 = "^2018.703"
pendulum = "^2.1.2"
plumbum = "^1.7.2"
prance = {version = "^0.21.8", extras = ["osv"]}
randomname = "^0.1.5"
requests = "^2.27.1"
yarl = "^1.7.2"

[tool.poetry.dev-dependencies]
invoke = "^1.6.0"
jsonpatch = "^1.32"
mypy = "^0.961"
pre-commit = "^2.16.0"
pytest = "^6.2.5"
requests-mock = "^1.9.3"
tbump = "^6.9.0"
types-PyYAML = "^6.0.8"
types-requests = "^2.27.30"
types-jsonschema = "^4.4.9"
types-python-dateutil = "^2.8.17"

[tool.poetry.scripts]
sinfonia-tier1 = "sinfonia.cli:tier1"
sinfonia-tier2 = "sinfonia.cli:tier2"

[tool.isort]
py_version = 37
profile = "black"

[tool.black]
target-version = ["py37"]

[tool.mypy]
check_untyped_defs = true
files = "src"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
